(1) Some rework in autogen.sh so that it is less verbose - and to provide an option "--fast" (which must be its first option) that uses (eg) "autoreconf" rather than "autoreconf -f", and that autoreconfs in each directory in parallel rather than sequentially. I can not at this stage be certain about the reliability of that mode, so it is not the default, and an autoconf.sh run without the --fast flag would always put things back in a safe state. Also try to put in a dependency in the CSL Makefile so that if parts of FOX where the source code is shared with CSL change that FOX gets re-build automatically. scripts/make.sh has been downgraded so that recompilation is always sequential even on multi-core platforms. This may slow building down but it makes log files a lot easier to read and avoids some risk of glitching! I may revisit this in the future if I can do so in ways I feel are fully safe.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5516 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ autogen.sh
@@ -8,7 +8,7 @@
 
 #
 # Usage:
-#     autogen.sh [options as for the configure script]
+#     autogen.sh [--fast] [options as for the configure script]
 #
 # If no arguments are passed the script will rebuild everything. If
 # --with-csl or --with-psl is specified only that section of the tree
@@ -16,6 +16,9 @@
 # and perhaps other sub-options will be checked to try to save reworking
 # directories that would not be used.
 #
+# "--fast" prevents it from using "force" options and so if things are
+# already up to date the steps should be faster, and it also does a lot in
+# parallel.
 
 # I used to have an option (enabled by default) to do the autoconf-ing
 # concurrently in all relevant source directories. However libtoolize
@@ -30,6 +33,15 @@ here=`cd \`dirname "$here"\` ; pwd -P`
 
 save=`pwd`
 cd $here
+
+if test "$1" = "--fast"
+then
+  fast="yes"
+  procids=""
+  shift
+else
+  fast="no"
+fi
 
 # On the Mac with macports the command needed here is glibtoolize rather than
 # libtoolize.
@@ -74,12 +86,19 @@ fi
 
 
 # I will re-process the top level first before any lower level
-rm -rf ltmain.sh config.cache autom4te.cache m4/libtool.m4 \
-       m4/lt-obsolete.m4 m4/ltoptions.m4 m4/ltsugar.m4 m4/ltversion.m4
 mkdir -p m4
-$LIBTOOLIZE --force --copy
-aclocal --force
-autoreconf -f -i -v
+if test "$fast" = "yes"
+then
+  $LIBTOOLIZE --copy
+  aclocal
+  autoreconf -i
+else
+  rm -rf ltmain.sh config.cache autom4te.cache m4/libtool.m4 \
+         m4/lt-obsolete.m4 m4/ltoptions.m4 m4/ltsugar.m4 m4/ltversion.m4
+  $LIBTOOLIZE --force --copy
+  aclocal --force
+  autoreconf -f -i
+fi
 
 # Here are the directories that I will always process...
 
@@ -127,31 +146,37 @@ case $a in
   ;;
 esac
 
-printf "@@@@ About to process $L @@@@\n"
-
-# At least as an experiment I will make the use of libtoolize run
-# sequentially rather than processing directories in parallel.
-
-s="$sequential"
-sequential=no
-
 for d in $L
 do
-  printf "\nlibtoolize in directory '%s'\n" $d
   if test -d $d
   then
     cd $d
-    printf "$LIBTOOLIZE --force --copy; aclocal --force\n"
-    mkdir -p m4
-    rm -rf ltmain.sh config.cache autom4te.cache m4/libtool.m4 \
-           m4/lt-obsolete.m4 m4/ltoptions.m4 m4/ltsugar.m4 m4/ltversion.m4
-    $LIBTOOLIZE --force --copy
-    aclocal --force
-    printf "autoreconf -f -i -v\n"
-    autoreconf -f -i -v
+    if test "$fast" = "yes"
+    then
+      mkdir -p m4
+      ( printf "$d: $LIBTOOLIZE/aclocal/autoreconf -i -v\n" && \
+        $LIBTOOLIZE --copy && \
+        aclocal && \
+        autoreconf -i ) &
+      procids="$procids $!"
+    else
+      mkdir -p m4
+      rm -rf ltmain.sh config.cache autom4te.cache m4/libtool.m4 \
+             m4/lt-obsolete.m4 m4/ltoptions.m4 m4/ltsugar.m4 m4/ltversion.m4
+      printf "$d: $LIBTOOLIZE --force --copy; aclocal --force\n"
+      $LIBTOOLIZE --force --copy
+      printf "$d: aclocal/autoreconf -f -i\n"
+      aclocal --force
+      autoreconf -f -i
+    fi
     cd $here
   fi
 done
+
+if test "$fast" = "yes"
+then
+  wait $procids
+fi
 
 scripts/resetall.sh
